CensusBuffer

Function to overlay point locations on ACS data at the block group level, draw concentric polygons (buffers) around point locations at user specified radial distance, and apportion block group populations to each buffer based on the percentage of spatial overlap.

Usage

CensusBuffer(
  buffer_vec,
  census_dat_proj,
  FacilityLocation_m,
  runTract = FALSE,
  tract_dat_proj = NULL,
  runCounty = FALSE,
  county_dat_proj = NULL,
  variableNames = NULL,
  EmissionData = NULL
)

Arguments

buffer_vec

A radial distance for a concentric buffer to be drawn around a point location in meters, or a vector of radial distances.

census_dat_proj

A spatial dataframe (sf object) of ACS data at the block group level, projected into a metric coordinate system. This should be the output of the PullCensusData function with a geography of "block group".

FacilityLocation_m

A spatial dataframe (sf object) of TRI facilities. This should be the output of the RSEISearch function, with returnLocation set to TRUE.

runTract

Logical. If true, adds tract level summaries for demographic variables for comparison. The user must supply ACS data summarized at the tract level. Defaults to False.

tract_dat_proj

A spatial dataframe (sf object) of ACS data at the tract level, projected into a metric coordinate system. This should be the output of the PullCensusData function with a geography of "tract".

runCounty

Logical. If true, adds county level summaries for demographic variables for comparison. The user must supply ACS data summarized at the county level. Defaults to False.

county_dat_proj

= A spatial dataframe (sf object) of ACS data at the countylevel, projected into a metric coordinate system. This should be the output of the PullCensusData function with a geography of "county".

variableNames

An optional vector of variable names, most likely the vector of named variable codes provided to the function PullCensusData.

EmissionData

An optional dataframe of emissions/RSEI score information if the user would like this merged into the demographic tabulation results. This should be the output of the RSEISearch function.

Value

A wide format dataframe. If EmissionsData are included, these data are merged into the final output so that each row is a chemical-facility combination, and the columns include the summed emissions and associated scores as well as the demography in each buffer. If EmissionsData is not included, the output of this function is a wide format dataframe with a row for each facility and in the columns the tabulated ACS variables within each buffer distance of that facility location. If runTract or runCounty are true, it also includes the county or tract totals for each variable for comparison. Counts of the number of TRI facilities within the buffer/tract/county are also included.

Details

This function is specifically designed to apportion ACS population data at the block group level to nearby TRI facilities in order to calculate the demography within a certain distance of those facilties. It relies on spatial functions in the sf package to take the outputs of the PullCensusData function (ACS data at the block group level, projected into a metric coordinate system) and the RSEISearch function (TRI Facility location data along with associated attributes, including emmissions) to determine the community attributes within a user specified distance of those facilities. Users can enter a vector of buffers (distance in meters) to do multiple calculations at once. For each radial distance provided, the data are intersected to determine the the proportion of each block groups that intersects the circular buffer. The population variables are then multiplied by that proportion to determine the number of people that are within the buffer, and this is summed for all the block groups that intersect each buffer to give a total population the buffer. This assumes that populations are distributed uniformly across block groups, which is incorrect but in the absence of finer scale distribution data provides a best estimate for the number of people that live within a given distance of each TRI facility.

This function is designed to be used with the output of the PullCensusData function, which pulls ACS data at desired geographical units, and the RSEISearch function, which produces a spatial dataframe of TRI reporting facilities as well as a dataframe of emissions data and RSEI scores for each chemical-facility combination.