CensusVariableSearch

This function allows users to search for variables by narrowing down the available options provided in the tidyCensus load_variables function.

Usage

CensusVariableSearch(
  Yr,
  dataset = "acs5",
  variableCode = NULL,
  geog_var = NULL,
  conceptSearch = NULL,
  labelSearch = NULL
)

Arguments

Yr

The end year of the survey you are pulling data from.

dataset

Survey type, which can be "acs5", "acs3", or "acs1". Defailts to "acs5".

variableCode

Character string or vector of character strings that can be used to search for a variable if you already know the code. For example, you can check if a code used from a previous survey is still valid.

geog_var

Optional string specifying the geography of the variable you are searching for, as this can help narrow results.

conceptSearch

Character string or vector of character strings representing a word or phrase found in the "concept" column of the variable table. Useful concept search terms include "Total population", "Median Household Income", "Race".

labelSearch

Character string or vector of character strings representing a word or phrase found in the "label" column of the variable table.

Value

A four column tibble of variables from the requested dataset, filtered by the search terms. If there are no matching search terms a warning will appear and an empty tibble will appear.

Details

The load_variables function allows users to select a dataset and year, but provides no additional means of filtering the thousands of variables that may be present. This function builds on the load_variables function to provide users a way to filter the variable table to reduce the number of variables shown.

Examples

CensusVariableSearch(Yr=2021, dataset="acs5", geog_var = "block group", conceptSearch = "Race")
#> # A tibble: 217 × 4
#>    name        label                                           concept geography
#>    <chr>       <chr>                                           <chr>   <chr>    
#>  1 B01002F_001 Estimate!!Median age --!!Total:                 MEDIAN… block gr…
#>  2 B01002F_002 Estimate!!Median age --!!Male                   MEDIAN… block gr…
#>  3 B01002F_003 Estimate!!Median age --!!Female                 MEDIAN… block gr…
#>  4 B01002G_001 Estimate!!Median age --!!Total:                 MEDIAN… block gr…
#>  5 B01002G_002 Estimate!!Median age --!!Male                   MEDIAN… block gr…
#>  6 B01002G_003 Estimate!!Median age --!!Female                 MEDIAN… block gr…
#>  7 B02001_001  Estimate!!Total:                                RACE    block gr…
#>  8 B02001_002  Estimate!!Total:!!White alone                   RACE    block gr…
#>  9 B02001_003  Estimate!!Total:!!Black or African American al… RACE    block gr…
#> 10 B02001_004  Estimate!!Total:!!American Indian and Alaska N… RACE    block gr…
#> # ℹ 207 more rows

CensusVariableSearch(Yr=2021, dataset="acs5", variableCode = "B02001_001")
#> # A tibble: 1 × 4
#>   name       label            concept geography  
#>   <chr>      <chr>            <chr>   <chr>      
#> 1 B02001_001 Estimate!!Total: RACE    block group

CensusVariableSearch(Yr=2021, dataset="acs5", geog_var = "block group", conceptSearch = "Education")
#> # A tibble: 119 × 4
#>    name       label                                            concept geography
#>    <chr>      <chr>                                            <chr>   <chr>    
#>  1 B14005_001 Estimate!!Total:                                 SEX BY… block gr…
#>  2 B14005_002 Estimate!!Total:!!Male:                          SEX BY… block gr…
#>  3 B14005_003 Estimate!!Total:!!Male:!!Enrolled in school:     SEX BY… block gr…
#>  4 B14005_004 Estimate!!Total:!!Male:!!Enrolled in school:!!E… SEX BY… block gr…
#>  5 B14005_005 Estimate!!Total:!!Male:!!Enrolled in school:!!U… SEX BY… block gr…
#>  6 B14005_006 Estimate!!Total:!!Male:!!Enrolled in school:!!N… SEX BY… block gr…
#>  7 B14005_007 Estimate!!Total:!!Male:!!Not enrolled in school: SEX BY… block gr…
#>  8 B14005_008 Estimate!!Total:!!Male:!!Not enrolled in school… SEX BY… block gr…
#>  9 B14005_009 Estimate!!Total:!!Male:!!Not enrolled in school… SEX BY… block gr…
#> 10 B14005_010 Estimate!!Total:!!Male:!!Not enrolled in school… SEX BY… block gr…
#> # ℹ 109 more rows