RSEISearch

This function allows users to access the EPAs' Risk-Screening Environmental Indicators (RSEI) Model database in order to obtain information on the total releases per facility per year, as well as the associated modeled risk and hazard scores. Users can search for specific time periods, facilities, or chemicals.

Usage

RSEISearch(
  DownloadRSEI = TRUE,
  RSEIpath = NULL,
  RSEIversion = "v2311",
  StartYr = NULL,
  EndYr = NULL,
  FacilityName = NULL,
  Facility_ID_vec = NULL,
  NAICS = NULL,
  State = NULL,
  County = NULL,
  City = NULL,
  ZipCode = NULL,
  CASNumbers = NULL,
  ChemicalNames = NULL,
  rsei_Chem_Numbers = NULL,
  returnLocation = FALSE,
  projectTo = NULL,
  wider = TRUE,
  Filename = "RSEIDataPull",
  SaveTo = NULL
)

Arguments

DownloadRSEI

Logical. If True, the function GetRSEIData is called to download the necessary RSEI files from the EPA website. Defaults to TRUE.

RSEIpath

Path specifying location where RSEI csv files will to be saved if DownloadRSEI is true, or the path specifying the location of existing RSEI csv files are stored.

RSEIversion

"v2311" or "v2310" to specify whether the RSEI data files are from the current version (23.11) or the previous version (23.10). Defaults to "v2311". Earlier versions of RSEI data are not supported.

StartYr

The first year of RSEI data to be selected.

EndYr

The last year of RSEI data to be selected. Note that the final year of data is 2021 in version 23.11 and 2020 in version 23.10.

FacilityName

Character string or vector of character strings specifying a name or partial name of a TRI facility. Capitalization does not matter.

Facility_ID_vec

Character string or vector of character strings specifying a TRI facility ID.

NAICS

6 digit numeric string representing a North American Industry Classification System (NAICS) code, or a vector of NAICS codes. Numbers provided to the NAICS argument are matched with those found in the column "Modeled NAICS". According to the RSEI data dictionary, a facility can report multiple NAICS codes. If more than one primary NAICS is reported by the facility, the most frequently reported primary NAICS for the most recent year is selected for the column "NAICS1", and the "Modeled NAICS" column is based on the NAICS1 but is modified based on additional information as necessary.

State

A character string or vector of character strings representing 2 letter state or territory codes.

County

Character string or vector of character strings specifying a name or partial name of a county. Capitalization does not matter.

City

Character string or vector of character strings specifying a name or partial name of a city. Capitalization does not matter.

ZipCode

5 digit numeric string, or a vector of 5 digit numeric strings, representing zip codes.

CASNumbers

A numerical string or vector of numerical strings representing a CAS Number. A CAS Registry Number is a numeric identifier that can contain up to 10 digits, separated by hyphens. However hyphens, dashes, and spaces are not allowed, so please collapse the standard CAS number format to a digital string.

ChemicalNames

Character string or vector of character strings specifying a name or partial name of a chemical. Capitalization does not matter.

rsei_Chem_Numbers

A number representing a unique idenifyer in the RSEI Chemicals table. Chemical Numbers range from 1 to almost 900 at this time.

returnLocation

Logical. If true returns a spatial dataframe (sf object) of the location of TRI facilities, including street address and latitude/longitude. A .csv file and Rdata object are also saved to the working directory. Defaults to FALSE.

projectTo

4 digit numeric string specifying a EPSG coordinate system. If included, the Lat/Long for each facility location are projected to the provided coordinate system in the dataframe containing TRI location information that is created if returnLocation is true.

wider

Logical. If true, data is pivoted so that for each chemical-facility pair, the information about releases and the associated hazard scores per year is in a wide format where there are columns for each year. If false, information about releases/hazards per year is in a long format, with one row for each facility-chemical-year combination. Defaults to TRUE.

Filename

Optional character string specifying name of saved output file. File name will be appended to current date. If not supplied defaults to "RSEIDataPull".

SaveTo

Optional path where final RSEI datafiles should be saved. If NULL, output is not saved.

Value

This function returns a dataframe containing the summed chemical releases per facility per year, as well as the associated risk/hazard scores for those releases, based on the search criteria provided. Information on each chemical and facility is also included in the dataframe.

There is also an option to return a second dataframe that contains the location information of the facilities returned in the search for use in spatial operations. This sf object contains the point geometry of the facilities along with their names, facility IDs, and addresses, projected into a desired coordinate system.

Details

The RSEI database includes data from the federally managed Toxic Release Inventory (TRI) program on over 25,000 facilities reporting hazardous emissions since the 1980s, as well as risk scores that account for the size of the chemical release, the fate and transport of the chemical through the environment, the size and location of the exposed population, and the chemical's toxicity. The aggregated version of these data is available for download on the RSEI website. The data are spread out over different tables, with crosswalks between them to link facilities and chemicals to emmissions data at the release level. This function allows users to either read in these tables if they have already been downloaded, or to download the necessary files, and then knits the tables together and filters them based on the search criteria provided to the function. These search criteria allow users to narrow these vast tables to specific years, chemicals, facility locations or types.

This function uses the function GetRSEIData to download and unzip the five csv files (Facility, Chemical, Submission, Releases, and Elements) necessary to calculate the summed releases per facility per year of each chemical. Users can also point the function to where these csv files have been saved. The function requires files to have the names they were downloaded with.

Examples

#Download RSEI tables and find yearly releases for all chemicals released by the Ritepack Inc.

RitePack <- RSEISearch(DownloadRSEI = TRUE, RSEIversion="v2310",  FacilityName = NULL,
Facility_ID_vec = "66024KMGBR52S15",  Filename = "RSEITest4", wider = FALSE)
#> Warning: URL 'https://gaftp.epa.gov/rsei/Older_Versions/Version2310_RY2020/Public_Release_Data/RSEIv2310_Public_Release_Data.zip': status was 'SSL peer certificate or SSH remote key was not OK'
#> Error in download.file("https://gaftp.epa.gov/rsei/Older_Versions/Version2310_RY2020/Public_Release_Data/RSEIv2310_Public_Release_Data.zip",     temp, quiet = FALSE): cannot open URL 'https://gaftp.epa.gov/rsei/Older_Versions/Version2310_RY2020/Public_Release_Data/RSEIv2310_Public_Release_Data.zip'